fix: 金额字段改为BigDecimal类型
问题：Reimburse实体中金额字段使用Double类型，可能存在精度问题
修复：
- 将money字段类型从Double改为BigDecimal
- 更新构造函数和getter/setter方法
影响：提高金额计算精度，避免浮点数精度问题

+++ src/main/java/fun/wsss/hmh/entity/Reimburse.java
@@ -5,6 +5,7 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -44,7 +45,7 @@ public class Reimburse {
      * 报销金额
      */
     @TableField("money")
-    private Double money;
+    private BigDecimal money;
 
     /**
      * 报销原因
@@ -110,7 +111,7 @@ public class Reimburse {
         super();
     }
 
-    public Reimburse(Integer id, Integer sqUserId, Integer spUserId, Integer cType, Double money, Integer status,
+    public Reimburse(Integer id, Integer sqUserId, Integer spUserId, Integer cType, BigDecimal money, Integer status,
                      String reason, String denyReason, Integer type, Integer typeId, String spName, String sqName,
                      Date createTime, Date endTime) {
         super();
@@ -146,7 +147,7 @@ public class Reimburse {
         this.cType = cType;
     }
 
-    public void setMoney(Double money) {
+    public void setMoney(BigDecimal money) {
         this.money = money;
     }
 
@@ -206,7 +207,7 @@ public class Reimburse {
         return cType;
     }
 
-    public Double getMoney() {
+    public BigDecimal getMoney() {
         return money;
     }
